Add mobile tasks overview card for dashboard welcome
