+++ frontend/src/features/dashboard/components/MobileTasksOverviewCard.module.css
@@ -1,0 +1,238 @@
+.card {
+  display: flex;
+  flex-direction: column;
+  gap: 18px;
+  padding: 20px;
+  color: rgba(255, 255, 255, 0.95);
+  background:
+    radial-gradient(140% 140% at 100% 0%, color-mix(in srgb, var(--brand, #fa3356) 18%, transparent) 0%, transparent 58%),
+    linear-gradient(180deg, rgba(255, 255, 255, 0.08) 0%, rgba(255, 255, 255, 0) 68%),
+    var(--bg2, #111213);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  border-radius: var(--radius-squircle, 20px);
+  box-shadow:
+    inset 0 0 0 1px rgba(255, 255, 255, 0.05),
+    0 12px 32px rgba(0, 0, 0, 0.35);
+  overflow: hidden;
+}
+
+@supports (color: color-mix(in srgb, white 50%, black 50%)) {
+  .card {
+    border-color: color-mix(in srgb, rgba(255, 255, 255, 0.08) 78%, transparent 22%);
+    box-shadow:
+      inset 0 0 0 1px color-mix(in srgb, rgba(255, 255, 255, 0.08) 55%, transparent 45%),
+      0 12px 32px rgba(0, 0, 0, 0.35);
+  }
+}
+
+.header {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 12px;
+}
+
+.titleWrap {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.title {
+  margin: 0;
+  font-size: 18px;
+  font-weight: 600;
+}
+
+.subtitle {
+  margin: 0;
+  font-size: 13px;
+  color: rgba(255, 255, 255, 0.65);
+}
+
+.viewAllButton {
+  border: none;
+  background: none;
+  color: rgba(255, 255, 255, 0.82);
+  font-size: 13px;
+  font-weight: 600;
+  text-decoration: underline;
+  padding: 6px 4px;
+  align-self: flex-start;
+}
+
+.viewAllButton:focus-visible {
+  outline: 2px solid var(--color-focus, var(--brand, #fa3356));
+  outline-offset: 2px;
+}
+
+.viewAllButton:hover {
+  color: rgba(255, 255, 255, 0.95);
+}
+
+.statRow {
+  display: flex;
+  gap: 10px;
+}
+
+.stat {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  padding: 12px 14px;
+  border-radius: 16px;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  background: rgba(255, 255, 255, 0.04);
+  min-height: 70px;
+}
+
+.statValue {
+  font-size: 20px;
+  font-weight: 600;
+}
+
+.statLabel {
+  font-size: 11px;
+  text-transform: uppercase;
+  letter-spacing: 0.04em;
+  color: rgba(255, 255, 255, 0.65);
+}
+
+.statOk {
+  border-color: color-mix(in srgb, var(--color-success, #4caf50) 48%, transparent);
+  background: color-mix(in srgb, var(--color-success, #4caf50) 14%, transparent);
+}
+
+.statWarn {
+  border-color: color-mix(in srgb, var(--color-warning, #ff9800) 48%, transparent);
+  background: color-mix(in srgb, var(--color-warning, #ff9800) 14%, transparent);
+}
+
+.statDanger {
+  border-color: color-mix(in srgb, var(--color-error, #ef5350) 52%, transparent);
+  background: color-mix(in srgb, var(--color-error, #ef5350) 18%, transparent);
+}
+
+.groups {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+}
+
+.group {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.groupHeader {
+  display: flex;
+  align-items: baseline;
+  justify-content: space-between;
+  gap: 8px;
+}
+
+.groupDay {
+  font-size: 13px;
+  font-weight: 600;
+  color: rgba(255, 255, 255, 0.85);
+}
+
+.groupCount {
+  font-size: 12px;
+  color: rgba(255, 255, 255, 0.6);
+}
+
+.items {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.item {
+  display: flex;
+  align-items: flex-start;
+  gap: 12px;
+}
+
+.itemDot {
+  width: 10px;
+  height: 10px;
+  border-radius: 50%;
+  margin-top: 6px;
+  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.35);
+}
+
+.itemBody {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.itemTitle {
+  font-size: 14px;
+  font-weight: 600;
+  color: rgba(255, 255, 255, 0.95);
+}
+
+.itemMeta {
+  font-size: 12px;
+  color: rgba(255, 255, 255, 0.62);
+}
+
+.empty {
+  font-size: 13px;
+  color: rgba(255, 255, 255, 0.6);
+  padding: 4px 0 2px;
+}
+
+.footerActions {
+  display: flex;
+  justify-content: stretch;
+}
+
+.primaryButton {
+  flex: 1;
+  padding: 12px 16px;
+  border-radius: 16px;
+  border: 1px solid color-mix(in srgb, var(--brand, #fa3356) 45%, transparent);
+  background: color-mix(in srgb, var(--brand, #fa3356) 18%, transparent);
+  color: color-mix(in srgb, var(--brand, #fa3356) 85%, white 15%);
+  font-size: 14px;
+  font-weight: 600;
+}
+
+.primaryButton:hover:not(:disabled) {
+  background: color-mix(in srgb, var(--brand, #fa3356) 26%, transparent);
+}
+
+.primaryButton:disabled {
+  opacity: 0.5;
+}
+
+.primaryButton:focus-visible {
+  outline: 2px solid var(--color-focus, var(--brand, #fa3356));
+  outline-offset: 2px;
+}
+
+@media (max-width: 480px) {
+  .card {
+    padding: 18px;
+  }
+
+  .statRow {
+    flex-wrap: wrap;
+  }
+
+  .stat {
+    flex: 1 1 45%;
+  }
+
+  .footerActions {
+    margin-top: 4px;
+  }
+}
